Added a sequence diagram, Added module that daemonizes the process
