+++ server/Makefile
@@ -1,10 +1,10 @@
 CC = g++
-SRC = $(wildcard *.cpp) $(wildcard ../lib/*.cpp) 
+SRC = $(wildcard *.cpp) $(wildcard *.c) $(wildcard ../lib/*.cpp) 
 IDIR = -I. -I.. -I$(HOME)/dev/include/
 LDDIRS = -L$(HOME)/dev/lib/
 LDFLAGS += -lboost_system -lcrypto -lssl -ltesseract -lcpprest
 
-langocr-server: 
+langocrd: 
 	$(CC) $(SRC) -o $@ $(IDIR) $(LDDIRS) $(LDFLAGS)
 
 .PHONY: clean
